Updated hardhat config
