+++ .gitignore
@@ -49,3 +49,6 @@ cmd/simulator/.simulator/*
 
 # goreleaser
 dist/
+
+# generator rpc file for e2e tests
+contract-examples/dynamic_rpc.json
